Incomplete, fix issue with base64 build parsing.

+++ Sources/SuiKit/Types/Enums/TransactionArgumentKind.swift
@@ -21,9 +21,9 @@ public enum TransactionArgumentKind: Codable, KeyProtocol {
     public func serialize(_ serializer: Serializer) throws {
         switch self {
         case .object:
-            try Serializer.u8(serializer, 0)
+            try Serializer.u8(serializer, UInt8(0))
         case .pure(let type):
-            try Serializer.u8(serializer, 1)
+            try Serializer.u8(serializer, UInt8(1))
             try Serializer.str(serializer, type)
         }
     }
