+++ Sources/SuiKit/Types/Structs/SuiBcs.swift
@@ -14,10 +14,10 @@ public indirect enum CallArg: Codable, KeyProtocol {
     public func serialize(_ serializer: Serializer) throws {
         switch self {
         case .pure(let pureCallArg):
-            try Serializer.u8(serializer, 0)
+            try Serializer.u8(serializer, UInt8(0))
             try Serializer._struct(serializer, value: pureCallArg)
         case .object(let objectArg):
-            try Serializer.u8(serializer, 1)
+            try Serializer.u8(serializer, UInt8(1))
             try Serializer._struct(serializer, value: objectArg)
         }
     }
